web/requirements.txt
asgiref==3.7.2
astroid==3.1.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dill==0.3.8
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==4.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyJWT==2.8.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
